+++ distribution/CHANGES.md
@@ -6,10 +6,11 @@ UCC Changelog
 
 **JAVA VERSION NOTE** This release requires Java 11 or later!
 
-Version 10.0.0 (released Feb dd, 2024)
+Version 10.0.0 (released Feb 21, 2024)
 --------------------------------------
- - new feature: "allocate" command for creating an allocation
+ - new feature: 'allocate' command for creating an allocation
  - 'shell': fix: OAuth authentication fails when refresh token is present, but expired
+ - 'shell: new feature: URL completion for UNICORE URLs
  - 'list-jobs': improved job list when using the "--long" option
  - 'issue-token': show subject when inspecting the token
  - 'exec': new option "--allocation" for running command in an existing allocation